fix rubocop breakages and separate bucket completion out
